So many changes.  bumped version to 0.2.1
This is the EOL commit for this repo probably.

I've removed all magic methods and the delegate support for the context
manager.  The way that logging factories are stored in the context
caused problems with python pickling.

This shows up as problems with executing code within a pyspark
enviroment like dataproc.

The only real solution is - don't keep objects around for more than you
need.  Python's threading.Lock objects inability to serialize will cause
grief.